requirements.txt
tldextract~=3.1.2
matplotlib~=3.3.2
seaborn~=0.11.1
nltk~=3.5
sklearn~=0.0
pyarrow
utils~=1.0.1
pandas~=1.1.5
scikit-learn~=0.23.2
pytorch
scipy~=1.5.2
xgboost~=1.5.0